Move income statement code to lib-common
