+++ frontend/src/citizen-frontend/income-statements/IncomeStatementEditor.tsx
@@ -9,6 +9,9 @@ import { combine, Loading, Result } from 'lib-common/api'
 import { IncomeStatementStatus } from 'lib-common/generated/api-types/incomestatement'
 import { IncomeStatementId } from 'lib-common/generated/api-types/shared'
 import { fromUuid } from 'lib-common/id-type'
+import { fromBody } from 'lib-common/income-statements/body'
+import * as Form from 'lib-common/income-statements/form'
+import { emptyIncomeStatementForm } from 'lib-common/income-statements/form'
 import LocalDate from 'lib-common/local-date'
 import {
   constantQuery,
@@ -28,9 +31,6 @@ import {
   incomeStatementStartDatesQuery,
   updateIncomeStatementMutation
 } from './queries'
-import { fromBody } from './types/body'
-import * as Form from './types/form'
-import { emptyIncomeStatementForm } from './types/form'
 
 interface EditorState {
   id: string | undefined
