+++ frontend/src/citizen-frontend/income-statements/IncomeStatementView.tsx
@@ -17,7 +17,11 @@ import { IncomeStatementId } from 'lib-common/generated/api-types/shared'
 import {
   collectAttachmentIds,
   toIncomeStatementAttachments
-} from 'lib-common/income-statements'
+} from 'lib-common/income-statements/attachments'
+import {
+  computeRequiredAttachments,
+  fromIncomeStatement
+} from 'lib-common/income-statements/form'
 import { useQueryResult } from 'lib-common/query'
 import { useIdRouteParam } from 'lib-common/useRouteParams'
 import HorizontalLine from 'lib-components/atoms/HorizontalLine'
@@ -46,7 +50,6 @@ import {
   incomeStatementQuery,
   updateSentIncomeStatementMutation
 } from './queries'
-import { computeRequiredAttachments, fromIncomeStatement } from './types/form'
 
 export default React.memo(function IncomeStatementView() {
   const incomeStatementId =
